Add zsh-autosuggestions and aliases to ksh setup
The commit adds zsh-autosuggestions and aliases to the ksh setup. The Brewfile is updated to include "zsh-autosuggestions". In the setup.sh script, the variable "$HOMEBREW_PATH" is changed to "$HOMEBREW_PREFIX". The script also appends a line to the .zshrc file to source the zsh-autosuggestions plugin and adds aliases from a separate file.