Add Docker setup to macOS bootstrap script
The Docker setup has been added to the macOS bootstrap script. This change includes the addition of a new `Brewfile` for Docker dependencies and a new `setup.sh` script that installs these dependencies using Homebrew. The Docker setup is now part of the overall development environment setup process in macOS.