Change installation of oh-my-zsh to only run if the directory does not exist.
The code changes in this commit modify the setup.sh file in the macos/shell/ksh directory. Previously, the script would unconditionally install oh-my-zsh using a curl command. However, with this change, the installation will only occur if the $HOME/.oh-my-zsh directory does not already exist. This ensures that oh-my-zsh is not reinstalled unnecessarily.

This change improves efficiency by avoiding redundant installations and reduces potential conflicts with existing configurations.