Change homebrew setup script to correctly export HOMEBREW_PREFIX and update PATH in .zshenv
The code changes the logic for exporting HOMEBREW_PREFIX and updating PATH in the .zshenv file. Previously, it was checking if the lines already existed in the .zshrc file, but now it checks in the .zshenv file. Additionally, it adds two new lines to source .zshenv and .zshrc files after running brew bundle.

This change ensures that HOMEBREW_PREFIX is correctly exported and PATH is updated when setting up Homebrew on macOS.