Install or upgrade Homebrew
This commit adds code to install or upgrade Homebrew. If Homebrew is not already installed, it will be installed using the installation script from the official Homebrew repository. If Homebrew is already installed, a message will be displayed indicating that it is already installed.

The code also includes a check for the system architecture (arm64 or x86_64), but this part of the code is currently commented out and does not have any effect.

No changes were made to the Brewfile or any other files in this commit.