Add .homebrew.zsh and source it in .zshrc
This commit adds a check for the existence of `.homebrew.zsh` file. If the file does not exist, it is created with the line `eval $(brew shellenv)`. Additionally, this commit checks if `.homebrew.zsh` is sourced in `.zshrc`, and if not, appends the line `source $HOME/.homebrew.zsh` to `.zshrc`. Finally, this commit upgrades Homebrew using `brew upgrade`.