Refactor setup.sh to update /etc/paths if necessary
The script now creates a temporary file called `paths` to store the paths. It checks if `$HOMEBREW_PREFIX/bin` and `$HOMEBREW_PREFIX/sbin` exist in `/etc/paths`, and if not, adds them to `paths`. The contents of `/etc/paths` are then appended to `paths`, removing any duplicate entries. If `/etc/paths` is different from the updated `uniq_paths`, it is replaced with the new file. Finally, the environment is updated using `$HOMEBREW_PREFIX/bin/brew shellenv`.