Add pwsh to /etc/shells if it is not already present
Previously, the setup script for PowerShell on macOS added pwsh to /etc/shells unconditionally. This change modifies the script to first check if pwsh is already in /etc/shells before adding it. If pwsh is already present, a message indicating that it is already in /etc/shells will be displayed. If pwsh is not present, it will be added using `brew --prefix` and `sudo tee -a`. This change improves the setup process by avoiding duplicate entries in /etc/shells.