Change 1Password setup script to sign in if config file doesn't exist
The setup.sh script for 1Password was modified to check if the config file exists before signing in. If the config file does not exist, the script will now execute the command to sign in using `op signin`. This change ensures that users are automatically signed in when running the setup script for 1Password.