Change 1Password setup script to sign in if config file doesn't exist
The setup.sh script for 1Password on macOS has been modified to include a check for the existence of the config file. If the file does not exist, the script will now prompt the user to sign in using their 1Password credentials. This change ensures that users are prompted to sign in when necessary, improving the overall user experience.