Change 1Password setup script to add reference in ~/.ssh/config
The 1Password setup script was modified to add a reference to 1Password in the ~/.ssh/config file if it doesn't already exist. The previous code used an incorrect format for the reference, which has been fixed. Now, the correct format is used and the reference is added using the `cat` command instead of `echo`. This change ensures that 1Password is properly configured in the ~/.ssh/config file.