Change bootstrap.sh to print a horizontal line before executing each setup script.
The previous implementation of bootstrap.sh did not include a visual separator between the execution of different setup scripts. This made it difficult to distinguish between the output of each script when running the bootstrap process.

To address this, I modified the code in bootstrap.sh to add a horizontal line before executing each setup script. This line is created by printing a series of dashes that spans the width of the terminal window.

By adding this visual separator, it is now easier for users to identify when one setup script ends and another begins during the bootstrap process.