Refactor setup_dir function to handle exclude argument
The setup_dir function in bootstrap.sh has been refactored to handle the exclude argument. Previously, the function did not account for this argument, but now it correctly checks if there is an exclude argument and stores it in a variable called "exclude". This change improves the functionality of the script by allowing users to specify directories or files to be excluded during setup.